added vertical offset and trigger level arrows

+++ src/Utility/LAB_Utility_Functions.h
@@ -7,7 +7,7 @@ namespace LABF
 {
   bool        is_equal          (double a, double b, double epsilon = 1e-9);
   bool        is_within_range   (double value, double min, double max);
-  int         normalize         (int input, int min_input, int max_input, int min_output, int max_output);
+  double      normalize         (double input, double min_input, double max_input, double min_output, double max_output);
   std::string get_now_timestamp ();
 };
